VideoEncoderService/requirements.txt
grpcio==1.66.1
grpcio-reflection==1.66.1
protobuf==5.28.0
python-dotenv==1.0.1
